Final version-1

+++ src/ent.c
@@ -103,7 +103,7 @@ static int getopt(int argc, char *argv[], char *opts)
 void ENT(                                                                                // int argc, char *argv[],
         unsigned char *input_array, size_t array_size, unsigned long binary,             // IN
         double *ent, double *chisq, double *mean, double *montepi, double *scc,          // OUT
-        double *chip, double *compRate, double *errorPct                                 // Additional
+        double *chip,  int *compRate, double *errorPct                                 // Additional
 )
 {
     size_t i;
